+++ pyccel/ast/class_defs.py
@@ -11,7 +11,8 @@ from pyccel.ast.builtin_methods.set_methods  import (SetAdd, SetClear, SetCopy, 
 from pyccel.ast.builtin_methods.list_methods import (ListAppend, ListInsert, ListPop,
                                                      ListClear, ListExtend, ListRemove,
                                                      ListCopy, ListSort)
-from pyccel.ast.builtin_methods.dict_methods  import DictPop, DictPopitem, DictGet, DictClear
+from pyccel.ast.builtin_methods.dict_methods  import DictPop, DictPopitem, DictGet, DictClear,
+                                                     DictSetDefault)
 
 from .builtins   import PythonImag, PythonReal, PythonConjugate
 from .core       import ClassDef, PyccelFunctionDef
@@ -174,10 +175,11 @@ SetClass = ClassDef('set',
 
 DictClass = ClassDef('dict',
         methods=[
+            PyccelFunctionDef('clear', func_class = DictClear),
             PyccelFunctionDef('get', func_class = DictGet),
             PyccelFunctionDef('pop', func_class = DictPop),
             PyccelFunctionDef('popitem', func_class = DictPopitem),
-            PyccelFunctionDef('clear', func_class = DictClear),
+            PyccelFunctionDef('setdefault', func_class = DictSetDefault),
         ])
 
 #=======================================================================================